+++ tests/PHPStan/Analyser/nsrt/bug-11276.php
@@ -41,18 +41,3 @@ function doFoo(int $i, int $j, $arr): void
 		assertType('non-empty-array', $arr);
 	}
 }
-
-function doBar($j, array $arr) {
-	$i = 1;
-	if (!array_key_exists($i, $arr)) {
-		if (array_key_exists($i, $arr) || array_key_exists($j, $arr)) {
-			assertType('non-empty-array<mixed~1, mixed>', $arr);
-		}
-	}
-
-	if (!array_key_exists($i, $arr)) {
-		if (array_key_exists($j, $arr) || array_key_exists($i, $arr)) {
-			assertType('non-empty-array<mixed~1, mixed>', $arr);
-		}
-	}
-}
